+++ src/main/java/com/renzzle/backend/domain/puzzle/community/api/request/AddCommunityPuzzleRequest.java
@@ -1,6 +1,6 @@
 package com.renzzle.backend.domain.puzzle.community.api.request;
 
-import com.renzzle.backend.domain.puzzle.shared.domain.Difficulty;
+import com.renzzle.backend.domain.puzzle.training.domain.Difficulty;
 import com.renzzle.backend.domain.puzzle.shared.domain.WinColor;
 import com.renzzle.backend.global.validation.ValidBoardString;
 import com.renzzle.backend.global.validation.ValidEnum;
