+++ src/main/java/com/renzzle/backend/domain/puzzle/community/domain/CommunityPuzzle.java
@@ -1,6 +1,6 @@
 package com.renzzle.backend.domain.puzzle.community.domain;
 
-import com.renzzle.backend.domain.puzzle.shared.domain.Difficulty;
+import com.renzzle.backend.domain.puzzle.training.domain.Difficulty;
 import com.renzzle.backend.domain.puzzle.shared.domain.WinColor;
 import com.renzzle.backend.domain.user.domain.UserEntity;
 import com.renzzle.backend.global.common.domain.Status;
