+++ src/main/java/com/renzzle/backend/domain/puzzle/training/dao/PackRepository.java
@@ -1,6 +1,6 @@
 package com.renzzle.backend.domain.puzzle.training.dao;
 
-import com.renzzle.backend.domain.puzzle.shared.domain.Difficulty;
+import com.renzzle.backend.domain.puzzle.training.domain.Difficulty;
 import com.renzzle.backend.domain.puzzle.training.domain.Pack;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
