+++ src/main/java/com/renzzle/backend/domain/puzzle/training/service/TrainingService.java
@@ -1,6 +1,6 @@
 package com.renzzle.backend.domain.puzzle.training.service;
 
-import com.renzzle.backend.domain.puzzle.shared.domain.Difficulty;
+import com.renzzle.backend.domain.puzzle.training.domain.Difficulty;
 import com.renzzle.backend.domain.puzzle.shared.domain.WinColor;
 import com.renzzle.backend.domain.puzzle.training.api.response.GetPackResponse;
 import com.renzzle.backend.domain.puzzle.training.api.response.GetTrainingPuzzleAnswerResponse;
